requirements.txt
shiny
posit-sdk>=0.9.0
